Convert to sbt-tut.
Le bien: it's helpful to intersperse the documentation with the
example code.  When a concept can't be introduced succinctly
inline, we've find incidental complexity to wring out of the
library.

Le mal: not DRY with examples.  It's not clear that the examples
directory matches a tutorial format in the long run.